+++ functions/read.py
@@ -2,9 +2,6 @@ from google.cloud import bigquery
 
 job_config = bigquery.QueryJobConfig()
 job_config.use_legacy_sql = False
-# client.query(query, job_config=job_config)
-
-# Initialize the BigQuery client with your project ID
 client = bigquery.Client(project="psp-portfolio-f1205", default_query_job_config=job_config)
 
 # Define your query