+++ webapp/src/index.ts
@@ -1,17 +1,19 @@
 import { Firestore } from "firebase/firestore";
 import { getTestInfo, getGeneralTestInfo } from "./db_interaction";
 import { update } from "./plotting";
-import type { DatasetStatus } from "./types";
-import { copyTextToClipboard, getSharelink, getTestName, initAdded } from "./browser_fxns";
-import { loader } from "./html_components";
+import { loadingStatus, type DatasetStatus } from "./types";
+import { getTestID, getSharelinkList, setTitle, updateStatus } from "./browser_fxns";
 import { initModal, setKnownTests } from "./modal";
 import { initFirebase } from "./firebase_init";
+import type uPlot from "uplot";
+import { setupEventListeners } from "./toolbar";
 
 declare global {
   var activeDatasets: DatasetStatus;
-  var test_name: string;
+  var test_id: string;
   var db: Firestore;
   var FIREBASE_APPCHECK_DEBUG_TOKEN: boolean | string | undefined;
+  var uplot: uPlot;
 }
 
 activeDatasets = {
@@ -24,26 +26,15 @@ initFirebase();
 initModal();
 
 async function main() {
-  document.getElementById("status")!.innerHTML = loader;
+  updateStatus(loadingStatus.LOADING);
   const [tests, default_url] = await getGeneralTestInfo();
   setKnownTests(tests, default_url);
-  test_name = getTestName(default_url);
-  initAdded();
-  const [datasets, name, test_article] = await getTestInfo();
+  test_id = getTestID(default_url);
+  getSharelinkList();
+  const [datasets, name, test_article, gse_article] = await getTestInfo();
   activeDatasets.all = datasets.sort((a, b) => a.localeCompare(b));
-  const titleElement = document.getElementById("title")!;
-  const sharelinkButton = document.getElementById("sharelinkButton")!;
-  const plotDiv = document.getElementById("plot")!;
-  const selectorDiv = document.getElementById("dataset-selector")!;
-  const tabTitle = document.getElementById("tabTitle")!;
-  titleElement.innerHTML = "PSP Data Viewer::" + test_article + "::" + name;
-  tabTitle.innerHTML = test_article + "::" + name;
-  sharelinkButton.style.opacity = "1";
-  sharelinkButton.addEventListener("click", async (e) => {
-    const sharelink: string = getSharelink();
-    copyTextToClipboard(sharelink);
-    console.log(sharelink);
-  });
+  setTitle(name, test_article, gse_article);
+  setupEventListeners();
   update();
 }